+++ webprac/src/main/java/dao/EducationDAO.java
@@ -8,7 +8,7 @@ import entity.Education;
 public interface EducationDAO {
 	void add(Education education) throws SQLException;
 	List<Education> getAll();
-	Education getById(int id);
+	Education getById(Long id);
 	void update(Education education);
 	void remove(Education education);
 }
